entry.s: Rework to avoid stomping registers.
From MBR we need DL and from PXE we need various other registers.  So,
we want to preserve them as much as possible prior to either the MBR or
PXE entry functions.  We were preserving most stuff before, but only
under the assumption that there was a working stack and in the MBR case
we can't be guaranteed that SS:SP is set up to anything usable.  So, now
we set up our own stack at 0:FC00 - but to do that we need at least one
register.  EAX seems a fine victim since we know it doesn't hold
anything and on the way back to BIOS for PXE it contains an error code -
so not something we'd need to preserve.