lokios.0: Always save/zero/restore ES and also record ES:BX for PXE.
In the MBR case we were already zeroing ES (the INT 13h API uses ES:BX
for the destination address) and on return to BIOS we actually weren't
restoring it - we were just lucky since bochs uses ES=0 anyways.  For
the second-stage loader we are going to need to do INT 15h which uses
ES:DI as a destination buffer and we need this for both MBR and PXE
cases - meaning that we are going to always stomp ES so we might as well
always save/restore it and set it up as 0 as part of init.