lokios.1: Link 16-bit sources in an intermediate output file.
In the future we are going to support 16-bit sources and then transition
into 64-bit mode to run the OS.  At a minimum we will have a 64-bit
trampoline to get into the OS.  64-bit code generation is different than
16-bit generation and normally the linker doesn't let you mix the two
(for obvious reasons).

So now what we have to do here is generate an intermediate 16-bit ELF
file.  Then use objcopy to change the intermediate file's architecture
to 64-bit (which just changes metadata) and then link it against all of
the 64-bit object files to make the final "64-bit" ELF file.

In the end it all works and we try to ignore how ugly it is.  Ripe for
future cleanup, no doubt.