lokios.1: Compile long_mode.s as a regular 16-bit real mode file.
We were doing this one specially because we were getting compile-time
error about rdmsr/wrmsr.  But that was because we had specified
-march=i386 which doesn't have those instructions.  We've now changed
the target architecture to "pentium" which does support those
instructions.  This doesn't change the code emitted (--32 sees to that)
but the target architecture is used to validate which instruction the
assembler is allowed to emit.