lokios.1: Turn it back into a simple i386 build process.
For the jump into long mode we were doing a whole bunch of useless crap
just so we could make a simple x86_64 .cc file build.  We can actually
build it as a 32-bit i386 .cc file instead and still do the long-mode
jump and everything will work.

That change means we don't have to force multiple architectures together
in the same file; this matters because the long jump is actually going
to be to a completely different binary loaded in high memory at address
2M and we won't ever be linking against it.  Since we won't link against
it that means we don't have to worry about mixed architectures.