lokios.0: Merge lokios.1 into lokios.0 and remove lokios.1.
It turns out we don't actually need the two stages to be built as
separate binaries now that we are more comfortable using linker scripts
and working with (around) the assembler and compiler architecture flags.

This commit simply merges the two directories; it works but we are going
to do a few optimizations and cleanup in the next few commits.