lokios.0: Have the bootsector branch directly into PXE or MBR entry points.
Everybody is linked together now, so we can move some of the PXE stuff
out of the bootsector into higher memory (freeing up space in the
bootsector is always good).  We also split the entry points into
separate files; in the future these will be responsible for either
PXE-loading the kernel or reading the rest of it off disk.  For now they
just branch into common code.