bootsector: Remove dispatcher.s and replace it with disk.s and a nice subroutine.
Instead of reading a sector-at-a-time, we now use the extended read BIOS
function that takes an LBA instead of that CHS crap.  It's also going to
read multiple sectors for us at a time.  Since this is now a function
instead of a one-time dispatcher, we can reuse it later when we need to
read the kernel from disk as well.