lokios.0: Get ready to fetch a kernel image to _kernel_base.
Currently we were switching in to long mode and then executing a local C
program that blanks the screen.  What we really plan to do is load a
kernel image at 0x00200000 and then branch to it.  Well, we don't have
the facility to load the kernel image yet (or a kernel image to load),
but we'd like to test that we can branch to the right address after
switching to long mode.

So, the PXE and MBR entry points do their jobs and now place an "image"
at _kernel_base which is simply a HLT instruction.  _enter_long_mode now
does its job and branches there.  bochs state shows us that we made it.

Next job: write a kernel and load it there...