PXE: Enough PXE support to get the first sector of the kernel and execute it.
This gets us 99% of the way there.  Figuring out what was failing (PXE
switches to protected mode then back to real mode, nuking our unreal
mode configuration) was a bit of a pain!  Now we just need to loop over
all packets in the file and then unload PXE before jumping into the
kernel to do this all properly.  Something for tomorrow...