bootsector: Save a few bytes by coding the upper 32-bits of the LBA to 0 in the struct.
Previously this was an optional argument if you had to use a 64-bit LBA.
For now we really don't need that, so it got changed to instead write
the value 0 to that field.  But if we pre-populate the field with 0 then
we can just throw the entire instruction away and save some space.