lokios.0: Factor common initialization code out into _common_init; enable A20 gate early.
It turns out that we want to enable the A20 gate before we start copying
kernel images around.  If the kernel image expands to be longer than 1M,
the A20 gate will cause quite a bit of mysterious trouble during the
bounce-buffer copy process!  So we now enable it properly before doing
the bounce-buffer copying.