MBR: Actually check the return code from _disk_read when loading the kernel.
We were checking this when loading the rest of lokios.0, but we weren't
checking it when reading the kernel.  If we tried to read past the end
of the disk in bochs, we were ignoring the error and "succeeding".
Really we should flag the problem and exit back to BIOS.