PXE: Execute a PXE shutdown sequence.
The PXE shutdown sequence is strange and mysterious.  I don't even know
if we really have to do any of this crap.  The Stop UNDI command seems
to be a problem on the MacBook Air with iPXE - if we execute that
command then when we jump to the kernel it sets the screen to blue,
prints the "Loki" text and then the laptop reboots within about 1
second.  I'm really not clear on what's going on there...  In any case,
this is resetting a lot of things and hopefully will be enough until a
real driver takes over the chip.