kassert: Rework the abort/assert stuff into kassert.
assert was a bad name because there is a standard library header named
assert.h and a standard library function called assert.  We don't really
want to collide with those.  aborts was also a weird name to choose when
what we are really doing is panicking with a string.  So, now you
basically just call panic() all the time unless you are external support
libraries that think they can call abort() which we still provide.