tmock: Do a better job of classifying test results and handling expected failures or unexpected passes.
The classification system is a bit wonky though: we can mark a test case
as expected to fail but there isn't a way to mark that this isn't
correct behavior.  I.e. we can't differentiate between these two cases:

1. A function that correctly calls abort() if you pass it a NULL pointer
   and we want to test that functionality in a test case that passes in
   a NULL pointer.  In this case, the test "failing" due to abort() is
   expected and code-under-test is also correct.
2. A function that doesn't implement everything yet and perhaps calls
   abort() on a subset of its input because that functionality is still
   missing.  In this case the test failure is expected but the
   code-under-test is incorrect.

So, we may need a richer set of descriptors to capture all these
possibilities and generate a more-correct report at the end of running
everything.