Makefile: Fix .cc -> .o recipe.
We generate .d header dependency files as part of the build.  The first
time we build something these dependencies don't exist, but they will
exist on future builds.  Normally, on future builds you don't need to
remake the file and so everything is fine.

However, if you do need to remake the .o file, the $^ syntax we switched
to recently will cause all of those extra header dependencies to be
listed on the g++ command line as well.  When you put headers on a g++
command line it thinks you want to build a precompiled header file, not
a .o file.  Bad things happen.

This undoes the change where we switched from $* to $^ syntax to fix this.