lokios.0: Include *timestamp.o in the bootsector data section.
At some point we started building bootsector.a instead of individual
object files and we link against the archive now.  We don't build or
include timestamp.o as part of bootsector.a because it is auto-generated
every time we link lokios.0.elf, not every time we build bootsector.a.
This caused timestamp.o to get picked up in the generic .data section
instead of the bootsector data section.  In that configuration, the MBR
stuff was actually printing a string from uninitialized RAM and luckily
not crashing out (because we haven't loaded anything other than the boot
sector yet!).

This commit moves timestamp.o's data section into the bootsector data
section so it's there when the entry code tries to print the banner.