tmock: Add silent mode flag.
When this mode flag is set, assertions won't print debug output.  This
is set by default if you are running a batch of test cases and cleared
if you are running an individual test.  That way we don't get spammed by
a bunch of error output, especially useful when we are testing failure
cases.