Makefile: If explicitly running 'make test', run all tests.
There isn't an easy way to specify the ordering of some sort of "clean
test results" target that removes the test results and a "run needed
tests" that tries to generate the missing test results.

We use a hack with MAKECMDGOALS to just nuke the test results directory
when we do 'make test'.