char_stream_test: Don't truncate char_stream's output!
We were using the number of characters returned from vasprintf to decide
how large to size the buffer for ss.printf().  This is a bad idea
because if ss.printf() is broken it might want to write more characters
than the buffer allows and the truncation to the length of the vasprintf
buffer would erroneously give us a test pass when the implementation was
actually broken.

Now we just double the string length to give ss.printf() lots of room to
screw up.