tmock: Allow for a richer classification of test outcomes.
We'd really like our test framework to do two things:

1. Tell if our tests are passing or failing; and,
2. Tell which of our tests aren't supposed to be passing or failing.

It's subtle, but it can be used to report on missing functionality.  For
instance, a test case could assert() on a NULL pointer.  In this case,
the sub-process exits with a signal and it looks as though the test case
has "failed".  However, we expect that failure and it SHOULD fail - if
it doesn't fail then the NULL pointer got through and maybe other bad
things could have happened.

So, we can now mark whether or not we expect a test case to fail and we
can also mark whether or not that test case should actually fail.  Put
these two flags together and you can find the functionality in your
system under test that still needs to be implemented.

Of course, this doesn't do things like find test cases that are failing
and should be failing but aren't failing for the right reasons - that's
left as an exercise for the reader.