kernel: Provide namespace kernel and move everything inside it.
This will help prevent name clashes with things like "assert" or
"printf".  We should have had it this way from the beginning...

As a part of doing this, kernel::kargs now has a weird name used in
entry.s.  I took the opportunity to also move the variable out of
entry.s into main.cc.