lokios.1: Enable SSE on entry to kernel.
This kind of sucks; some standard headers return double values and these
seem to require SSE registers.  At the same time, g++ is generating code
in the middle of our C++ soures that uses SSE registers to move data
around.

If we use -mno-sse, then it doesn't generate those SSE register moves,
but then the compiler complains about the standard headers that use
double.

For now we just turn it on; in the future we are going to have to
completely replace libc with our own subset that we care about.