sbrk: Add set_sbrk() function to allow sbrk resetting.
Early on we are going to borrow from the sbrk pool to pre-populate the
page pool.  However, we don't want those temporary pages to remain in
the page pool because they can confuse our simple malloc/free
implementations because they are below the sbrk line.  We'd like to
reclaim those pages into the sbrk pool and the only way to do that is to
reduce the sbrk line - which set_sbrk allows.