lokios.0.ld: Mark 4K pages with entry PAT0 like we have for the 2M pages.
The PAT bit moved around between the 4K and the 2M entries.  We
previously just copied the 2M entry format to define the 4K pages, but
it turns out that we copied a must-be-1 bit in the 2M entry that maps to
the PAT bit in the 4K entry.  Luckily this maps to the same permissions
in the page attribute table, but clear them for consistency with the 2M
pages.