page_table: Don't track the page_table pages with a vector<>.
A vector<> is limited in size to a single page.  Worse, if we start
doing dynamic mappings and unmappings then it will be hard to find the
corresponding page entries in the vector<> to erase() them and thus
release them back to the page pool.

It turns out that the page table itself is a pretty handy-dandy data
structure that happens to point at every single page that we own.  Just
use it directly for cleanup purposes...  D'oh.