page_table: Program the PAT MSR in a way so that PTE PAT bits become "don't care".
The data structures in the x86 page table are a bit inconsistent
depending on if you are looking at a 4K, 2M or 1G page.  In particular,
the PAT bit moves around.  For the 2M and 1G pages, the bit where the
PAT bit lives in a 4K page must be set to 1 (and the PAT bit is
somewhere else).

It would be nice to treat all these types the same and always just set
that bit to 1.  This means we'd also be setting PAT=1 for the 4K pages
and 0 everywhere (since we can't force the 2M/1G bit to 1 in a 4K page
entry since that bit is used by something else).

Since we've gone and configure the PAT MSR in a way that means that the
PAT bit in the page tables is now a "don't care" bit, we can now treat
all these page table data structures the same and ignore the fact that
for 4K pages we will set PAT=1 and for 2M/1G pages we will set PAT=0.