page_table: Add page table test and fix first bug.
This bug causes a memory access past the end of a page table page.
Depending on what's in that next memory location it could be benign or
it could cause a SEGV.  Unfortunately, this means that the test case
that would expose the bug is unreliable and we don't have a way in tmock
to mark a test as unreliable.  So that bisect always works, I'm
therefore putting the bug fix in the same commit as the unittest instead
of doing the normal 2-step procedure where we introduce the unittest
exposing the bug then fix them in the next commit.