long_mode: Remove most of the TLS support.
We don't create a TLS anymore and just place a short frame at the top of
the stack and point the MSR_FS_BASE register at the start of that frame.
We do need a minimally-functional TLS because g++ does stack-stomp
checking and it gets the signature out of an OS-provided field at
FS:0x28.  So, we create an unpopulated 64-byte stack frame there so that
any early g++ stack-stomp operations won't trigger a SEGV until we can
finally get the kernel_task spawned.