entry.s: Fix completely broken interrupt-handling code.
This was supposed to be replacing the iretq stack frame with values from
the new thread context, but instead it was loading them from the thread
context and saving them back into the thread context.  Oops.

We also need to take care of the CS/SS registers - although, they are
always going to be constants in our kernel so we might make an
optimization in the future.