main: Perform cpu initialization as part of main().
It doesn't actually make sense to do a bunch of extra initialization
steps in the init/init2() methods - we should do as little as possible
to get into a usable state, which means we need page_alloc() to work and
we need a task/thread so that TLS works.  Everything else can come
later.