klist: Make klink constexpr work.
g++7 complains about this one because an integer cast to a pointer value
isn't a constexpr.  We sneakily use a union so that we can do the
constexpr initialization with a uintptr_t and later just use the pointer
field instead.  I manually verified that the constexpr constructor
works at compile-time.