init: Reorder the init sequence.
We actually want to initialize the CPU before starting the first task
because task state includes segment-register values.  This was working
before because the first entries in the lokios.0 GDT match what we are
using in the kernel so the selector values are the same, but we don't
actually want to have to rely on that.

So, initialize the CPU first, then initialize the task and have it start
via a bounce-to-main function that wraps main() and does halt() if
main() ever returns.