interrupts: Enable and test interrupts.
We now have everything masked out, so we can safely turn on the IF bit.
This allows us to then use the local APIC to send ourselves an
interrupt.  Since there is a bit of latency waiting for the interrupt,
we also had to get some delay code working via the pmtimer.