init.cc: Hoist exception-handling initialization up a little earlier.
It turns out that we don't actually need the init_array stuff to run
before we register the exception-handling support.  Perhaps due to
switching to a newer compiler or perhaps I was just mistaken or had a
bug in the older code.  Anyways, doing this should allows exceptions to
work for global/static constructors now too.