build: Allow test objects to be built with different flags than kernel objects.
In a future commit we're going to build the kernel with -mno-sse and
-mno-sse2 to prevent the compiler from inserting SSE instructions (and
thus to prevent the kernel from stomping XMM registers).  However, this
breaks the build any time we include a header that uses a float or
double (even if said functions are never invoked).  We do want to be
able to use all of the standard C/C++ libraries in our test binaries,
however (since things like set<> can really simplify some test cases) so
we don't want to restrict what can be included there.  That means the
test binaries need different compile-time flags...