slab: Convert the slab to not use an object size template argument.
This makes the slab type generic - for instance we can now store a bunch
of slab* pointers in an array even if the slabs have different object
sizes.  This could be useful for a slab-based malloc() implementation
which might take the log2 of the size parameter as an index into an
array of slabs to find the slab that most-closely matches the requested
object size.