smp: Refactor initialize sequence, pull code out of main, provide a scheduler.
I couldn't really break this down into a sequence of working commits.
The basic idea here is that the BSP will now initialize the kernel task
and spawn the main thread (which won't be scheduled yet), create a cpu*
and _thread_jump() into a second-stage function that will then go and do
all of the ACPI and interrupt controller initialization that main() was
doing.  Once that is out of the way, it configures the local APIC and
then invokes the new mini-scheduler.

The scheduler will see the previously-spawned main thread queued up and
will make it runnable on the BSP.  The main thread will then start the
other AP CPUs and they will start scheduling threads as well.