schedule: Rework into a true scheduler object that is timer- and interrupt-aware.
This incorporates a timer wheel, a local work queue and a remote work
queue into a single scheduler object.  We also add methods to schedule
work immediately on a local or remote cpu as well as schedule deferred
work on the local cpu.  Scheduling deferred work on a remote cpu is not
supported.