main: Look for a 'TSTi' ACPI table entry to decide whether or not to exit.
qemu allows adding acpi table entries using the -acpi command line
argument.  We use this to tell lokios whether or not we are in
integration-test mode or regular mode.  In integration-test mode we want
lokios to trigger a qemu exit with a successful exit code so that our
automated integration tests will know that we passed.  In regular mode
we are maybe running it by hand and want to do things like watch the
tickers tick and don't want it to exit until we kill it manually.