lokios.1: Add ability to link in driver archives.
We define a driver to have a name (i.e. "e1000") and provide a symbol
named the driver name suffixed with _driver (i.e. "e1000_driver").  That
symbol is used to pull in the driver archive; typically it will be a
kernel::pci::driver global that knows how to match against a PCI device
and instantiate a kernel::pci::dev to drive it.