mp_tables: Add code to find the MP tables which can be used to figure out PCI interrupt routing.
In ACPI, to figure out the PCI interrupt routing you need to use the AML
interpreter (this is the only way to support things like hot-swap I
guess).  We don't have an AML interpreter and probably won't for a very
long time.  The MP tables also provide routing information for PCI
devices, but only for the ones that were present when boot firmware was
executing.  That should be enough for our purposes, so we may use this
in the future for non-MSI devices.