entry.s: Define MSI-X interrupt entry points for interrupts 128-255.
These interrupt slots will look up their corresponding work_entry in the
cpu struct, queue it up on the cpu's local work queue, mask the MSI-X
interrupt in the device's MSI-X table and EOI the interrupt in the local
APIC.  We try to make these as lightweight as possible so that
interrupts aren't very expensive.  There's probably still a lot of room
for optimization.