sbrk: Freeze sbrk after early init time.
We use sbrk() to get some big chunks of memory for the
exception-handling code.  After that, we don't want to use sbrk() and in
fact we want malloc() to return NULL if the requested size is larger
than what we can satisfy (a page) - __cxa_allocate_exception is the only
use for this stuff and returning NULL will cause the exception-handling
code to fall back to an internal pool (presumably allocated previously
via sbrk) and allocate the exception from there.