schedule: Refactor timer_entry out of wqe_entry, use a kdlist.
We are going to want to cancel timers in the future and to do this in
O(1) time we require a circular doubly-linked list.  That means we need
another pointer in wqe type but we don't want that space taken up in
regular work items (such as a DSR) - that would require rewriting the
ISR assembly that manipulates the work queues (yuck).