kernel: Stop using identity-mapped pages.
This changes page_alloc() to return pages in the 0xFFFF800000000000
virtual address range instead of a pointer that is the same as the
physical address.  While it makes things more complex right now, it is
going to simplify things in the future when we start using the first
half of the address space to hold vm-alloc'd blocks and they will not be
identity-mapped or even physically contiguous.