mm: Use 4K mappings for pmap() and pmap_range().
In the future we are going to map all of usable RAM (as indicated via
the E820 tables) into the 0xFFFF800000000000 area.  It turns out that
the ACPI tables are all in areas of RAM that aren't marked as usable
(which is good, otherwise we'd stomp them) but if we try to randomly map
ACPI pages using 2M hugepages we will end up colliding with edge 4K
pages that get mapped when we map all of usable RAM.

For iomap we should probably also avoid doing 2M pages (especially if a
device has less than 2M of register space) so it picks up this change
automatically.