page: Perform a second stage page init that maps all usable RAM.
The first stage maps the first 8M of RAM (or whatever we got from the
bootloader).  Then we create the kernel_task and page table and at that
point we can then actually start mapping the rest of RAM.  We iterate
over the remainder of the usable pages in the E820 list map them and
link them up onto the free_pages_list as we go.  This means that as we
iterate over the pages they become available for use by the system (say,
to provide pages to the page table itself to map even more pages in a
really really large system).

We also try to use 2M hugepages wherever possible, but we do map smaller
regions using 4K pages if necessary.