task: When cloning the bootloader page tables, ignore all virtual addresses in the bottom half.
The bootloader needs to be able to access both low and high memory in
order to set up the kernel image in memory.  However, the kernel doesn't
need to use low-memory/identity pages at all.  With this change we no
longer map the bottom-half addresses and do everything in virtual space.