slab: Fix broken assertion for non-power-of-2 slab elements.
The assert was testing that the address of a slab element was a multiple
of its size.  That's true for objects with a power-of-2 size but it
isn't true in general; the general rule is that the offset of the
element from the start of the slab page should be a multiple of its
size.

This commit corrects the assert.