mm: Implement pmap_range in terms of mmap instead of pmap.
pmap() only does 4K mappings while mmap() is smart and can do 2M or 1G
mappings.  This results in saving ~77 page table pages when mapping the
MCFG space on the MacBook Air (bringing it back down to 23 total page
table pages which is in line with the 24 pages used under QEMU).