arp: Respond to arp requests.
We need to do this if we want to be able to telnet in on real hardware.
Apparently qemu doesn't do arp to find us (after all, it already knows
all the hardware addresses) but on a real network if we don't respond
then a telnet command won't be able to send packets to us.