mm: Get fpage to validate that the vaddr will work as a paddr.
There are restrictions in the x86 page tables on what can be used as a
paddr - no CPU actually supports a full 64-bit paddr.  To avoid
accidentally generating an illegal paddr we check that the vaddr could
map to a valid paddr before actually stuffing it in a fake page table or
something.