hash_table: Dynamically resize the hash table to maintain a load factor between 2-6.
When the load factor exceeds 6 we double the table size to bring it down
to 3.

When the load factor falls below 2 we halve the table size to bring it
up to 4 (with the exception that the table size never drops below
PAGE_SIZE).

This provides hysterisis so that we don't get into a pathological
situation where we are constantly resizing the table after manipulating
only a couple elements.