virtio_net: Support legacty virtio-net devices.
The qemu pxe bios doesn't recognize non-legacy virtio devices, so in
order to do pxe testing under qemu we have to omit the argument that
disables legacy support.  That's okay; the chip still has modern support
anyways and the lokios driver works with it just fine if we match
against the legacy device id.