lokios.0.ld: Provide a .bss section.
It seems like this isn't handled properly by objcopy if one doesn't
exist - ld will add one but put it at the end of the image (which makes
sense), but objcopy won't actually instantiate it!  The image will be
truncated instead of zero-filled at the end meaning that we'll end up
with garbage values in our .bss globals instead of zeroes.