bootsector: Move the global descriptor table into the bootsector.
This makes life simpler later if we just let the bootsector initialize
this.  We previously had a GDT for unreal mode and a different one for
the switch to long mode.  We merge these into a single GDT and also add
entries for 32-bit protected mode which we will use shortly.