bootsector: Invoke the mode32 entry point instead of the 16-bit ones.
With this commit we switch over to using all the protected-mode 32-bit
C++ code in the mode32 library instead of all the assembly code floating
in the top-level lokios.0/ directory.  This also slightly changes the
calling convention into the kernel, so we modify the lokios.1 bsp entry
point to handle it.