lokios.1.ld: Properly check that our length is a multiple of the sector size.
This was previously checking that we end on a page boundary which was a
bit strict.  It was also checking that the end of the image was on the
boundary rather than the length of the image.  We care about the length
of the image because this is what lokios.0 is going to have to fetch; in
a future commit the image won't start on a page boundary and it would
screw up the calculation of the number of sectors.