lokios.0: Remove m32 prefixes except where necessary.
The use of the m32_ prefix throughout mode32 was because we had function
names that could collide with similarly-named ones in the 16-bit part of
the tree.  But we've long since thrown away all the 16-bit code except
for the bootsector, so we don't need those prefixes to avoid name
clashes any longer.