elf: Add dumb elf image handling.
If the elf image is meant to be loaded at 2M and all of the sections are
fully populated in the image (i.e. none that need zero-filling by the
loader) then we can load the elf image in place and execute it.

Coincidentally, this is exactly how the lokios.1.elf image is crafted.
I wonder how that happened?