pxe: Support closing the TFTP stream early.
We currently don't read the entire elf image when processing an elf
stream.  We'd still like to be able to jump into the kernel when we have
as much as we currently care about.  Add support for the TFTP_CLOSE
command (it's not really clear what this does or if it's necessary) to
allow closing the stream early.