elf: Smarter elf handling.
We are now able to parse enough of the file to ensure that we read in
all of its contents.  Previously, we were relying on the LOKI header to
find the number of sectors of kernel image to read, but that would omit
all of the sections after the kernel, including that symbol table we are
after.