mode32: Disable interrupts after doing disk IO.
The INT 13h call to read sectors from disk seems to have a bad habit of
returning with interrupts enabled (at least, on my MacBook Air).  Since
this is the first time I've actually tested MBR on real hardware, it
went unnoticed.  When an interrupt fires and we are still in the
bootloader, the system will blow up and immediately reset.  Keep
everything cleared until we get into the kernel which can actually
handle interrupts properly.