mode32: Pass the address of the end of the kernel image as a karg.
The kernel currently sets _sbrk after at the end of the .ld file.  But
this doesn't take into account all of the elf headers and sections that
the linker will put after that address.  We want to not stomp all over
those because we will do symbol lookups in them in the future.  Pass the
actual end of the image as a karg so that the kernel can make better
life choices.