net: Change rx api such that rx handler need to return instructions on how to complete frame processing.
Currently we will auto-delete based on the contents of a frame's flags
field.  However, if the frame is to be handled by the client, there are
a number of scenarios that could happen where the frame has actually
been deleted by the time we return to the interface:: code that would
check the flags field:

1. The frame is handled by an external client inline on the callback and
   the external client decides to free it.  The external client may be a
   TCP client and not be aware at all of things like frame boundaries.
2. The frame is deferred to a queue for asynchronous processing and that
   queue is handled by another CPU.  Then we race against that other CPU
   processing and deleting the frame vs. us checking the flags field.

So basically it is unsafe to look at any of the rx_page fields after
we've passed it to a client unless the stack tells us specifically that
it is safe to do so.