net: Add client_payload and client_len fields to rx_page.
These are for the client's use after consuming a packet; the client may
need to process the page in steps and these values can be used as a type
of iterator to advance over the packet's remaining data.  Since typical
MTUs will be much less than the 4K rx_page size, in practice this isn't
going to hurt us much.  It also 16-byte aligns the payload which could
possibly be a performance improvement when processing the packet.