endian: Fix confusing choice of swap function.
to_*_endian and from_*_endian both do exactly the same thing: swap the
type without any actual knowledge of whether or not the input was little
or big endian.  They are both provided so that the client code can be
more clear about which type is the input and which type is the output.

The be/le_uint types were using them backwards which is confusing to the
reader, although technically not incorrect.