interface: Add half-implemented tcp_connect() method and tcp_pipe_test unittest.
This version of tcp_connect is broken in a number of ways:

1. Passes NULL for the link-layer header.
2. Doesn't arp the destination IP address.
3. Returns a tcp::socket*.

A real implementation wouldn't be able to return a tcp::socket* because
for that we need a proper link-layer header and that would require
arping the address - an asynchronous process.  A real implementation
would have to go to the eth::interface to kick off an arp lookup, then
notify somebody in net::interface that we have the link-layer address,
the net::interface could instantiate the tcp::socket, send the SYN and
then notify the client that the socket now exists.  Or possibly wait for
TCP_ESTABLISHED before notifying.

So why do this?  Because it's half-implemented enough to work as an
active connection initiator in a unittest that uses finteraces - which
don't have any link-layer headers!  So we can basically start to write a
test that is almost a mini integration test by hooking up an active
tcp::socket to a passive one.  No cables or ethernet drivers required!