socket: Add send_op type to track transmit state.
Client-side operations are just arrays of alps without any type of
TCP/IP headers - the client simply streams data through the stack.  This
struct records the client-side operation.  It also records whether or
not a SYN precedes or a FIN succeeds the client payload.

All of this information is necessary to implement a generic TCP
retransmit queue.