socket: Use better values for snd_mss.
The initial snd_mss value will be set to 536 bytes - the largest MSS
that any host *must* accept.  If an MSS option is present when we later
receive a SYN, then we'll use that MSS instead of the small initial one.

In all cases, we'll also validate the send MSS against what the
interface is actually able to transmit.