socket: Don't allow options to push a packet size beyond the send MSS.
The sum of options + payload must not exceed the send MSS.  If the send
window is small, then we can probably safely insert options without
exceeding the send MSS.  But if the send window is large, then the
current code could have inserted the options and then appended a full
MSS worth of data which, together, would have exceeded the limit.

In practice it would probably never happen since we only insert options
on a SYN packet and our SYN packets typically don't carry data.
Still...