socket: Handle rx segment overlap.
When we receive a segment, in the event that this was part of a
retransmit by the remote guy it may overlap with data that we've
already received.  So we have to give only the new data to the client
rather than the entire packet.

This change also makes it so we can simplify this processing quite a
bit, so this bit of code has been refactored.