socket: Handle crossed-SYNs and self-connect.
We do this by adding some sub-states to TCP_SYN_SENT.  IMO the TCP spec
gets this one wrong.  The spec states that when you get a crossed SYN,
both ends transition out of SYN_SENT into SYN_RECVD and then transmit
SYN/ACKs to each other.

But the SYN part of the SYN/ACK is a retranmit - we've already sent the
original SYN to each other and don't need to send it again!  All each
side needs to do is ACK the other guy's SYN and wait for the other guy
to ACK our original SYN and we are done.