net: Introduce net::interface::tcp_unlink() method.
This unlinks a tcp::socket from the interface's hash table so that it
will no longer be able to process inbound packets, which also frees up
the port for reuse by a new socket.  However, it doesn't destroy the
tcp::socket, allowing the client to continue doing things like draining
the receive queue and later destroying the socket when ready.