socket: Upon entering TCP_CLOSED, only tcp_unlink the socket and invoke socket_reset.
The client created the socket; it should be the client's job to also
delete it.  This unlinks the socket from the interface's socket hash
table so that it won't receive any more packets.  The client can process
the remainder of the receive queue in due course and eventually must
invoke tcp_delete on the socket.