Makefile: Allow specifying the optimization level via OPT_LEVEL=-On on the make command line.
In general, the kernel doesn't compile if the optimization level is too
low.  We lose things like built-in strcmps and elimination of function
calls that are known at compile time to never be invoked.  This leads to
link errors and other issues.

However, it can be useful when building a single failing unittest via:

make tests/failing_test OPT_LEVEL=

to disable optimizations (or lower them to something like -O1) and get
easier-to-understand stack traces in gdb.