tcp: Hoist passive SYN validation out of the socket state machine.
The LISTEN state is a bit unique - it isn't really a state.  We have a
listener object watching for rx packets on that port, but we don't
really want to instantiate a full tcp::socket object if the SYN packet
is malformed or formatted in a way that is just going to cause us to
immediately reset the connection and transition to CLOSED.

So, we get the top-level tcp rx handler to pre-validate the SYN and send
the generic responses if it is malformed and only instantiate a
tcp::socket if we're really going to accept the SYN and transition out
of LISTEN immediately.