socket: Transition based on FIN before handing rx payload to the client.
This simplifies the process_payload_synchronized() method a bit and will
allow us to push the FIN checking up the call chain in the next commit.
It also could be useful for the client to check in their
socket_readable() method to know that we're not going to get any more
data after this point, although presumably we'll still need a
FIN-received callout for the case where we get FIN with no payload so
socket_readable() doesn't get reinvoked.